functions/authenticate_user_v2/requirements.txt
psycopg2-binary==2.9.10
cryptography==42.0.5
pyotp==2.9.0 